+++ NEWS.md
@@ -1,4 +1,6 @@
 # SDMtune 1.0.1.9000
+Main changes:
+* The function `thinData` accept now a matrix or a dataframe with several columns, useful if the users has information related to the coordinates that doesn't want to lose with the thinning procedure.
 
 # SDMtune 1.0.1
 Bug fix: